chore!: better emmylua grammar support
